requirements.txt
# Core Dependencies
streamlit==1.31.1
pandas==2.1.4
numpy==1.24.3

# Data Collection
pykrx==1.0.46
yfinance==0.2.36

# Machine Learning
scikit-learn==1.3.2
xgboost==2.0.3

# Technical Analysis
ta==0.11.0
pandas-ta==0.3.14b0

# Visualization
plotly==5.18.0
matplotlib==3.8.2
seaborn==0.13.1

# Utilities
python-dotenv==1.0.1
requests==2.31.0
beautifulsoup4==4.12.3

# Performance
joblib==1.3.2
tqdm==4.66.1

# Date/Time
python-dateutil==2.8.2
